adbsolute factorisation: split and moved into the
"proper" place
