+++ src/NumField/NfAbs.jl
@@ -28,4 +28,5 @@ include("NfAbs/NormRelation.jl")
 include("NfAbs/PolyFact.jl")
 include("NfAbs/MPolyGcd.jl")
 include("NfAbs/MPolyFactor.jl")
+include("NfAbs/MPolyAbsFact.jl")
 include("NfAbs/ConjugatesNS.jl")
