+++ test/NumField/NfAbs.jl
@@ -1,4 +1,5 @@
 @testset "NumField/NfAbs" begin
   include("NfAbs/ConjugatesNS.jl")
   include("NfAbs/MPolyGcd.jl")
+  include("NfAbs/MPolyAbsFact.jl")
 end
